pip_requirements.txt
--extra-index-url https://maven.earthdata.nasa.gov/repository/python-repo/simple
# This file should contain requirements to be installed via Pip.
harmony-service-lib==0.0.7
pydap ~= 3.2
WebOb ~= 1.8.6
webobtoolkit ~= 0.2.4
requests ~= 2.23.0
